feat!: enforce that interface names do not begin with an I
The current convention in NextRequest is to prefix typescript interface
names with `I`. This PR proposes enforcing the opposite: that interface
names are not prefixed with `I`.

I personally am not used to interface names prefixed with `I`, so it
is extra noise that is not providing me with useful information. That's
not a technical argument though, and if others find it useful then I'm
happy to keep it.