feat!: enforce that interface names do not begin with an I (#4)
The current convention in NextRequest is to prefix typescript interface
names with `I`. This PR proposes enforcing the opposite: that interface
names are not prefixed with `I`.

I personally am not used to interface names prefixed with `I`, so it
is extra noise that is not providing me with useful information. That's
not a technical argument though, and if others find it useful then I'm
happy to keep it.

Most of what I've read recommends not using I:
- https://dev.to/mscamargo/why-you-should-avoid-using-the-i-prefix-for-interfaces-in-typescript-43gd
- https://passionfordev.com/you-dont-need-to-prefix-interfaces-in-typescript-with-i/
- The typescript codebase's own style guide: https://github.com/Microsoft/TypeScript/wiki/Coding-guidelines#names